workspaces/Waiver_Dashboard/backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
python-multipart
pytz
apscheduler
supabase==2.28.3
PyJWT
passlib[bcrypt]
bcrypt==4.0.1
